pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openvegas"
version = "0.1.0"
description = "Terminal Arcade for Developers — mint, wager, and redeem $V for AI inference"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "Pillow>=10.0",
    "qrcode[pil]>=7.4",
    "textual>=0.50",
    "httpx>=0.27",
    "supabase>=2.0",
    "websockets>=15,<16",
    "python-jose[cryptography]>=3.3",
    "openai>=1.0",
    "anthropic>=0.30",
    "google-generativeai>=0.5",
    "stripe==14.3.0",
    "redis[hiredis]>=5.0",
    "celery>=5.3",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "asyncpg>=0.29",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[project.scripts]
openvegas = "openvegas.cli:cli"

[tool.setuptools.packages.find]
include = ["openvegas*", "server*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["third_party_external", ".git", ".history", ".pytest_cache"]
asyncio_default_fixture_loop_scope = "function"
